+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Scripts.hs
@@ -12,6 +12,7 @@
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -72,12 +73,14 @@ import Cardano.Ledger.Binary.Coders (
 import Cardano.Ledger.Core
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
-  Mem,
   MemoBytes (Memo),
   Memoized (..),
+  byteCountMemoBytes,
   getMemoRawType,
   mkMemoBytes,
-  mkMemoized,
+  mkMemoizedEra,
+  packMemoBytesM,
+  unpackMemoBytesM,
  )
 import Cardano.Ledger.Shelley.Scripts (
   ShelleyEraScript (..),
@@ -87,14 +90,13 @@ import Cardano.Ledger.Shelley.Scripts (
   pattern RequireMOf,
   pattern RequireSignature,
  )
-import Data.MemPack
-
 import Cardano.Slotting.Slot (SlotNo (..))
 import Control.DeepSeq (NFData (..))
 import Data.Aeson (ToJSON (..), (.=))
 import qualified Data.Aeson as Aeson
 import Data.ByteString.Lazy (fromStrict)
 import Data.ByteString.Short (fromShort)
+import Data.MemPack
 import Data.Sequence.Strict as Seq (StrictSeq (Empty, (:<|)))
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set (Set, member)
@@ -208,24 +210,28 @@ instance Era era => DecCBOR (Annotator (TimelockRaw era)) where
 -- They rely on memoBytes, and TimelockRaw to memoize each constructor of Timelock
 -- =================================================================
 
-newtype Timelock era = TimelockConstr (MemoBytes TimelockRaw era)
+newtype Timelock era = TimelockConstr (MemoBytes (TimelockRaw era))
   deriving (Eq, Generic)
-  deriving newtype (ToCBOR, NoThunks, NFData, SafeToHash, MemPack)
+  deriving newtype (ToCBOR, NFData, SafeToHash)
 
+instance Era era => MemPack (Timelock era) where
+  packedByteCount (TimelockConstr mb) = byteCountMemoBytes mb
+  packM (TimelockConstr mb) = packMemoBytesM mb
+  unpackM = TimelockConstr <$> unpackMemoBytesM (eraProtVerLow @era)
+
+instance Era era => NoThunks (Timelock era)
 instance Era era => EncCBOR (Timelock era)
 
-instance Memoized Timelock where
-  type RawType Timelock = TimelockRaw
+instance Memoized (Timelock era) where
+  type RawType (Timelock era) = TimelockRaw era
 
 deriving instance Show (Timelock era)
 
 instance EqRaw (Timelock era) where
   eqRaw = eqTimelockRaw
 
-deriving via
-  Mem TimelockRaw era
-  instance
-    Era era => DecCBOR (Annotator (Timelock era))
+instance Era era => DecCBOR (Annotator (Timelock era)) where
+  decCBOR = fmap TimelockConstr <$> decCBOR
 
 -- | Since Timelock scripts are a strictly backwards compatible extension of
 -- MultiSig scripts, we can use the same 'scriptPrefixTag' tag here as we did
@@ -286,39 +292,39 @@ pattern RequireTimeStart mslot <- (getTimeStart -> Just mslot)
   , RequireTimeStart
   #-}
 
-mkRequireSignatureTimelock :: Era era => KeyHash 'Witness -> Timelock era
-mkRequireSignatureTimelock = mkMemoized . Signature
-getRequireSignatureTimelock :: Era era => Timelock era -> Maybe (KeyHash 'Witness)
+mkRequireSignatureTimelock :: forall era. Era era => KeyHash 'Witness -> Timelock era
+mkRequireSignatureTimelock = mkMemoizedEra @era . Signature
+getRequireSignatureTimelock :: Timelock era -> Maybe (KeyHash 'Witness)
 getRequireSignatureTimelock (TimelockConstr (Memo (Signature kh) _)) = Just kh
 getRequireSignatureTimelock _ = Nothing
 
-mkRequireAllOfTimelock :: Era era => StrictSeq (Timelock era) -> Timelock era
-mkRequireAllOfTimelock = mkMemoized . AllOf
-getRequireAllOfTimelock :: Era era => Timelock era -> Maybe (StrictSeq (Timelock era))
+mkRequireAllOfTimelock :: forall era. Era era => StrictSeq (Timelock era) -> Timelock era
+mkRequireAllOfTimelock = mkMemoizedEra @era . AllOf
+getRequireAllOfTimelock :: Timelock era -> Maybe (StrictSeq (Timelock era))
 getRequireAllOfTimelock (TimelockConstr (Memo (AllOf ms) _)) = Just ms
 getRequireAllOfTimelock _ = Nothing
 
-mkRequireAnyOfTimelock :: Era era => StrictSeq (Timelock era) -> Timelock era
-mkRequireAnyOfTimelock = mkMemoized . AnyOf
-getRequireAnyOfTimelock :: Era era => Timelock era -> Maybe (StrictSeq (Timelock era))
+mkRequireAnyOfTimelock :: forall era. Era era => StrictSeq (Timelock era) -> Timelock era
+mkRequireAnyOfTimelock = mkMemoizedEra @era . AnyOf
+getRequireAnyOfTimelock :: Timelock era -> Maybe (StrictSeq (Timelock era))
 getRequireAnyOfTimelock (TimelockConstr (Memo (AnyOf ms) _)) = Just ms
 getRequireAnyOfTimelock _ = Nothing
 
-mkRequireMOfTimelock :: Era era => Int -> StrictSeq (Timelock era) -> Timelock era
-mkRequireMOfTimelock n = mkMemoized . MOfN n
-getRequireMOfTimelock :: Era era => Timelock era -> Maybe (Int, (StrictSeq (Timelock era)))
+mkRequireMOfTimelock :: forall era. Era era => Int -> StrictSeq (Timelock era) -> Timelock era
+mkRequireMOfTimelock n = mkMemoizedEra @era . MOfN n
+getRequireMOfTimelock :: Timelock era -> Maybe (Int, StrictSeq (Timelock era))
 getRequireMOfTimelock (TimelockConstr (Memo (MOfN n ms) _)) = Just (n, ms)
 getRequireMOfTimelock _ = Nothing
 
-mkTimeStartTimelock :: Era era => SlotNo -> Timelock era
-mkTimeStartTimelock = mkMemoized . TimeStart
-getTimeStartTimelock :: Era era => Timelock era -> Maybe SlotNo
+mkTimeStartTimelock :: forall era. Era era => SlotNo -> Timelock era
+mkTimeStartTimelock = mkMemoizedEra @era . TimeStart
+getTimeStartTimelock :: Timelock era -> Maybe SlotNo
 getTimeStartTimelock (TimelockConstr (Memo (TimeStart mslot) _)) = Just mslot
 getTimeStartTimelock _ = Nothing
 
-mkTimeExpireTimelock :: Era era => SlotNo -> Timelock era
-mkTimeExpireTimelock = mkMemoized . TimeExpire
-getTimeExpireTimelock :: Era era => Timelock era -> Maybe SlotNo
+mkTimeExpireTimelock :: forall era. Era era => SlotNo -> Timelock era
+mkTimeExpireTimelock = mkMemoizedEra @era . TimeExpire
+getTimeExpireTimelock :: Timelock era -> Maybe SlotNo
 getTimeExpireTimelock (TimelockConstr (Memo (TimeExpire mslot) _)) = Just mslot
 getTimeExpireTimelock _ = Nothing