+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,10 +2,30 @@
 
 ## 1.17.0.0
 
+* Remove `era` type parameter from `MemoBytes` type
+* Remove `Era era` constraint from:
+  * `Memo` pattern
+  * `decodeMemoBytes`
+  * `DecCBOR (Annotator (MemoBytes t))` instance
+  * `DecCBOR (MemoBytes t)` instance
+  * `memoBytes`
+  * `mkMemoized`
+  * `lensMemoRawType`
+  * `Data` pattern
+  * `dataToBinaryData`
+* Introduce `mkMemoizedEra` and `memoBytesEra`
+* Add `Version` parameter to:
+  * `memoBytes`
+  * `mkMemoized`
+  * `lensMemoRawType`
+* Remove `era` type parameter from `Mem` type
+* Reduce the kind of `MemoHashIndex` type family parameter to a concrete type
+* Reduce the kind of `RawType` type to a concrete type
+* Add `byteCountMemoBytes`, `packMemoBytesM` and `unpackMemoBytesM` to `MemoBytes` module
 * Require `MemPack` instance for `TxOut` and `CompactForm (Value era)` for `EraTxOut` type class.
 * Add `decodeMemoBytes`
 * Add `MemPack` instance for `CompactAddr`, `TxIx`, `TxId`, `TxIn`, `CompactForm Coin`,
-  `KeyHash`, `ScriptHash`, `Credential`, `SafeHash`, `MemoBytes`, `Plutus`, `PlutusBinary`, `BinaryData` and `Datum`
+  `KeyHash`, `ScriptHash`, `Credential`, `SafeHash`, `Plutus`, `PlutusBinary`, `BinaryData` and `Datum`
 * Add `DecShareCBOR` instance for `TxIn`
 * Add `fromCborRigorousBothAddr`
 * Add `SlotNo32` and use it in `Ptr` definition